Renamed Class48->Decimator
Renamed Class48->Decimator and moved to the sound package.
