+++ src/main/java/org/runejs/client/Class40_Sub12_Sub1.java
@@ -1,5 +1,7 @@
 package org.runejs.client;
 
+import org.runejs.client.sound.Decimator;
+
 public class Class40_Sub12_Sub1 extends Class40_Sub12 {
     public byte[] aByteArray2889;
     public int anInt2890;
@@ -13,14 +15,14 @@ public class Class40_Sub12_Sub1 extends Class40_Sub12 {
         anInt2891 = arg3;
     }
 
-    public Class40_Sub12_Sub1 method875(Class48 arg0) {
-        aByteArray2889 = arg0.method926(aByteArray2889, false);
-        anInt2892 = arg0.method923(anInt2892, 4);
+    public Class40_Sub12_Sub1 method875(Decimator arg0) {
+        aByteArray2889 = arg0.resample(aByteArray2889);
+        anInt2892 = arg0.scaleRate(anInt2892);
         if(anInt2890 == anInt2891)
-            anInt2890 = anInt2891 = arg0.method921(anInt2890, 54);
+            anInt2890 = anInt2891 = arg0.scalePosition(anInt2890);
         else {
-            anInt2890 = arg0.method921(anInt2890, 54);
-            anInt2891 = arg0.method921(anInt2891, 54);
+            anInt2890 = arg0.scalePosition(anInt2890);
+            anInt2891 = arg0.scalePosition(anInt2891);
             if(anInt2890 == anInt2891)
                 anInt2890--;
         }
