+++ src/main/java/org/runejs/client/Class40_Sub3.java
@@ -24,6 +24,7 @@ import org.runejs.client.scene.Scene;
 import org.runejs.client.scene.tile.FloorDecoration;
 import org.runejs.client.scene.tile.WallDecoration;
 import org.runejs.client.scene.util.CollisionMap;
+import org.runejs.client.sound.Decimator;
 
 public class Class40_Sub3 extends Node {
     public static ImageRGB[] aClass40_Sub5_Sub14_Sub4Array2019;
@@ -112,7 +113,7 @@ public class Class40_Sub3 extends Node {
         } else if (Class40_Sub5_Sub6.loadingPercent == 45) {
             MovedStatics.method446(Main.signlink, 0, !VertexNormal.lowMemory, (byte) 30);
             Class49.aClass40_Sub9_Sub1_1152 = MouseHandler.method1003(Main.signlink, MouseHandler.gameCanvas, arg0 ^ 0x74be);
-            Class55.aClass48_1289 = new Class48(22050, CollisionMap.anInt141);
+            Class55.aDecimator_1289 = new Decimator(22050, CollisionMap.anInt141);
             Class40_Sub5_Sub6.loadingPercent = 50;
             Native.currentLoadingText = English.preparedSoundEngine;
             MovedStatics.anInt1607 = 35;
