+++ src/main/java/org/runejs/client/Class59.java
@@ -31,7 +31,7 @@ public class Class59 {
     public static void dropClient() {
         if(SceneCluster.idleLogout > 0) {
             // Instant logout
-            Class48.logout();
+            MovedStatics.logout();
         } else {
             // Connection lost
             MovedStatics.processGameStatus(40);
